# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documintation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Получение JWT-токен, используя заданные учетные данные",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "example@exapmle.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "Password1@#"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0Ojg4NzYvYXBpL3VzZXJzL2xvZ2luIiwiaWF0IjoxNjk0MjMzNDI5LCJleHAiOjE2OTQyMzcwMjksIm5iZiI6MTY5NDIzMzQyOSwianRpIjoid011VEtxeHZXOVlxQU5pUCIsInN1YiI6IjUxIiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.fT5hk6B0wgFKffQ29YWoOZ-gv45itznc0EzSJDbyuWU"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/auth/me": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Получить аунтефицированного пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "FirstName"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "LastName"
                                        },
                                        "role": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "gender": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Some address"
                                        },
                                        "telephone": {
                                            "type": "string",
                                            "example": "+375259530534"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "example@example.com"
                                        },
                                        "email_verified_at": {
                                            "oneOf": [
                                                {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-09-05T10:22:25.000000Z"
                                                },
                                                {
                                                    "type": "nullable",
                                                    "example": "null"
                                                }
                                            ]
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-09-05T10:22:25.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-09-05T10:22:25.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/auth/refresh": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Получить обновленный токен",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0Ojg4NzYvYXBpL3VzZXJzL2xvZ2luIiwiaWF0IjoxNjk0MjMzNDI5LCJleHAiOjE2OTQyMzcwMjksIm5iZiI6MTY5NDIzMzQyOSwianRpIjoid011VEtxeHZXOVlxQU5pUCIsInN1YiI6IjUxIiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.fT5hk6B0wgFKffQ29YWoOZ-gv45itznc0EzSJDbyuWU"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/auth/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Выйти из системы(сделать токен не действительным)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/auth/wishlist": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Список понравившихся продуктов.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": "50.25"
                                            },
                                            "image": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "path": {
                                                        "type": "string",
                                                        "example": "/tmp/b472bb4d6f8f0f642b25a0466a4a2895.png"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://via.placeholder.com/255x310.png/001100?text=similique"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}